Now DescribableModelTest needs Jenkins.getInstance(), so distinction doesn't make sense
